requirements.txt
aiogram>=3.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
loguru>=0.7.0
vkbottle
setuptools<82
pydantic<2
requests